Rewrite param processing to be more clearly defined.
