+++ packages/babel-plugin-transform-es2015-destructuring/test/fixtures/destructuring/parameters/expected.js
@@ -21,15 +21,15 @@ console.log(unpackObject({
 }));
 
 var unpackArray = function (_ref3, _ref4) {
-  var _ref6 = babelHelpers.slicedToArray(_ref3, 3),
-      a = _ref6[0],
-      b = _ref6[1],
-      c = _ref6[2];
+  var _ref5 = babelHelpers.slicedToArray(_ref3, 3),
+      a = _ref5[0],
+      b = _ref5[1],
+      c = _ref5[2];
 
-  var _ref5 = babelHelpers.slicedToArray(_ref4, 3),
-      x = _ref5[0],
-      y = _ref5[1],
-      z = _ref5[2];
+  var _ref6 = babelHelpers.slicedToArray(_ref4, 3),
+      x = _ref6[0],
+      y = _ref6[1],
+      z = _ref6[2];
 
   return a + b + c;
 };
