+++ packages/babel-plugin-transform-es2015-parameters/src/index.js
@@ -1,32 +1,23 @@
-import type { NodePath } from "babel-traverse";
-import { visitors } from "babel-traverse";
-
-import * as destructuring from "./destructuring";
-import * as def from "./default";
-import * as rest from "./rest";
+import convertFunctionParams from "./params";
+import convertFunctionRest from "./rest";
 
 export default function() {
   return {
-    visitor: visitors.merge([
-      {
-        ArrowFunctionExpression(path) {
-          // In some conversion cases, it may have already been converted to a function while this callback
-          // was queued up.
-          if (!path.isArrowFunctionExpression()) return;
+    visitor: {
+      Function(path) {
+        if (
+          path.isArrowFunctionExpression() &&
+          path
+            .get("params")
+            .some(param => param.isRestElement() || param.isAssignmentPattern())
+        ) {
+          // default/rest visitors require access to `arguments`, so it cannot be an arrow
+          path.arrowFunctionToExpression();
+        }
 
-          // default/rest visitors require access to `arguments`
-          const params: Array<NodePath> = path.get("params");
-          for (const param of params) {
-            if (param.isRestElement() || param.isAssignmentPattern()) {
-              path.arrowFunctionToExpression();
-              break;
-            }
-          }
-        },
+        convertFunctionRest(path);
+        convertFunctionParams(path);
       },
-      destructuring.visitor,
-      rest.visitor,
-      def.visitor,
-    ]),
+    },
   };
 }
