+++ packages/babel-plugin-transform-es2015-parameters/test/fixtures/parameters/default-rest-mix/actual.js
@@ -1,0 +1,8 @@
+function fn(
+  a1,
+  a2 = 4,
+  {a3, a4},
+  a5,
+  {a6, a7} = {}) {
+
+}
